requirements.txt
pyyaml~=6.0.2
pandas~=2.0.3
matplotlib~=3.7.5